Switching from bigint to smallint.

+++ migrations/2024-11-21-195004_add_report_count/up.sql
@@ -1,12 +1,12 @@
 -- Adding report_count and unresolved_report_count
 -- to the post and comment aggregate tables
 ALTER TABLE post_aggregates
-    ADD COLUMN report_count bigint NOT NULL DEFAULT 0,
-    ADD COLUMN unresolved_report_count bigint NOT NULL DEFAULT 0;
+    ADD COLUMN report_count smallint NOT NULL DEFAULT 0,
+    ADD COLUMN unresolved_report_count smallint NOT NULL DEFAULT 0;
 
 ALTER TABLE comment_aggregates
-    ADD COLUMN report_count bigint NOT NULL DEFAULT 0,
-    ADD COLUMN unresolved_report_count bigint NOT NULL DEFAULT 0;
+    ADD COLUMN report_count smallint NOT NULL DEFAULT 0,
+    ADD COLUMN unresolved_report_count smallint NOT NULL DEFAULT 0;
 
 -- Update the historical counts
 -- Posts
